pyproject.toml
[project]
name = "librarymcp"
version = "0.1.0"
description = "Library backend server to serve MCP tool invokation"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.0",
    "uvicorn>=0.41.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pytest>=9.0.2",
    "mcp[cli]>=1.26.0",
    "httpx>=0.28.1",
    "pytest-httpx>=0.35.0",
    "python-jose[cryptography]>=3.5.0",
    "bcrypt>=5.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
